streamlit/requirements.txt
branca==0.3.1
cursor==1.3.5
folium==0.13.0
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.3
parquet-tools==0.2.11
psycopg2==2.9.5
pyarrow==9.0.0
python-dotenv==0.21.0
requests==2.28.1
seaborn==0.12.0
SQLAlchemy==1.4.42
streamlit==1.15.1
matplotlib==3.5.2
